выполнила задание 5.16. Функции возвращаются
